pyproject.toml
# https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html
[tool.black]
line-length = 120

# https://beta.ruff.rs/docs/settings/
[tool.ruff]
line-length = 120
# https://beta.ruff.rs/docs/rules/
select = ["E", "W", "F"]
ignore = ["F401", "F601"]
# Exclude a variety of commonly ignored directories.
respect-gitignore = true
ignore-init-module-imports = true

[project]
name = "python-yolo"
dynamic = ["version"]

dependencies = [
    "numpy>=1.23.0",                                 # temporary patch for compat errors https://github.com/ultralytics/yolov5/actions/runs/9538130424/job/26286956354
    "matplotlib>=3.3.0",
    "opencv-python>=4.6.0",
    "pillow>=7.1.2",
    "pyyaml>=5.3.1",
    "requests>=2.23.0",
    "scipy>=1.4.1",
    "torch>=1.8.0",
    "torch>=1.8.0,!=2.4.0; sys_platform == 'win32'", # Windows CPU errors w/ 2.4.0 https://github.com/ultralytics/ultralytics/issues/15049
    "torchvision>=0.9.0",
    "tqdm>=4.64.0",                                  # progress bars
    "psutil",                                        # system utilization
    "py-cpuinfo",                                    # display CPU info
    "pandas>=1.1.4",
    "seaborn>=0.11.0",                               # plotting
    "ultralytics-thop>=2.0.0",                       # FLOPs computation https://github.com/ultralytics/thop
]
